requirements.txt
fastapi
uvicorn
openai
anthropic
python-dotenv
pydantic
pydantic[email]
IPython
resend
typing
pytest
numpy
faiss-cpu
